refactor: integrate net_api
