+++ src/rpc/auth_layer.rs
@@ -4,7 +4,7 @@
 use crate::auth::{verify_token, JWT_IDENTIFIER};
 use crate::key_management::KeyStore;
 use crate::rpc::{
-    auth_api, beacon_api, chain_api, common_api, gas_api, mpool_api, state_api, sync_api,
+    auth_api, beacon_api, chain_api, common_api, gas_api, mpool_api, net_api, state_api, sync_api,
     wallet_api, CANCEL_METHOD_NAME,
 };
 use crate::rpc_api::*;
