+++ src/rpc/mod.rs
@@ -11,7 +11,7 @@ pub mod common_api;
 mod eth_api;
 pub mod gas_api;
 pub mod mpool_api;
-mod net_api;
+pub mod net_api;
 mod node_api;
 pub mod state_api;
 pub mod sync_api;
@@ -33,7 +33,7 @@ use crate::rpc::auth_layer::AuthLayer;
 use crate::rpc::channel::RpcModule as FilRpcModule;
 pub use crate::rpc::channel::CANCEL_METHOD_NAME;
 use crate::rpc::state_api::*;
-use crate::rpc_api::{eth_api::*, net_api::*, node_api::NODE_STATUS};
+use crate::rpc_api::{eth_api::*, node_api::NODE_STATUS};
 
 use fvm_ipld_blockstore::Blockstore;
 use hyper::server::conn::AddrStream;
